Refactor the code to use more OOP
